Move UserTypeValidationError to exceptions
